+++ src/Misc/Hooks.Timers.cpp
@@ -2,16 +2,20 @@
 #include <GameOptionsClass.h>
 #include <FPSCounter.h>
 #include <SessionClass.h>
+#include <Phobos.h>
 
 int oldValue;
 DEFINE_HOOK(0x6D4B50, Print_Timer_On_Tactical_Start, 0x6)
 {
+	if (!Phobos::Config::RealTimeTimers)
+		return 0;
+
 	REF_STACK(int, value, STACK_OFFSET(0, 0x4));
 	oldValue = value;
 
 	if (false && !SessionClass::IsMultiplayer()) // TODO change when custom game speed gets merged
 	{
-		value = std::max(value / (int)(std::max((double)FPSCounter::CurrentFrameRate, 1.0) / 15.0), 1);
+		value = (int)((double)value / (std::max((double)FPSCounter::CurrentFrameRate, 1.0) / 15.0));
 		return 0;
 	}
 
@@ -46,6 +50,9 @@ DEFINE_HOOK(0x6D4B50, Print_Timer_On_Tactical_Start, 0x6)
 
 DEFINE_HOOK(0x6D4C68, Print_Timer_On_Tactical_End, 0x8)
 {
+	if (!Phobos::Config::RealTimeTimers)
+		return 0;
+
 	REF_STACK(int, value, STACK_OFFSET(0x654, 0x4));
 	value = oldValue;
 	return 0;