+++ docs/Whats-New.md
@@ -277,6 +277,7 @@ New:
 - Attached particle system for animations (by Starkku)
 - Removal of hardcoded AA & Gattling weapon selection restrictions (by Starkku)
 - Projectile `SubjectToLand/Water` (by Starkku)
+- Real time timers (by Morton)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)